[grid-ui] Implementing more compact layout of stereotypes and visualizing node DOWN status with light-gray background
